feats: added sd-card task and sd card channels, configured peripheral uart gps, tested complete system, test has prints, changed time in messages to float, changed gpios to switches, flight computer esp32s3 uses channels, errors commented out in anticipation for system health system

+++ code/flight-computer-lib/src/device/bmp280.rs
@@ -50,7 +50,7 @@ where
             altitude,
             pressure,
             temperature, 
-            timestamp: Time::new::<microsecond>(Instant::now().as_micros()),
+            timestamp: Time::new::<microsecond>(Instant::now().as_micros() as f64),
         })
     }
 }
